# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: vanthang.thophu@gmail.com
  version: 1.0.11
servers:
  - url: https://backend-nextjs.onrender.com/
  - url: http://localhost:3002
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - auth
      summary: Register user
      description: Register user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@gmail.com
                password:
                  type: string
                  example: 123456789Kha@
              required:
                - email
                - password
          application/xml:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/auth/me:
    get:
      tags:
        - auth
      summary: Get my info of user
      description: Get details info of user
      operationId: getAuthMe
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - auth
      summary: Update info by me
      description: Update info by me
      operationId: updateAuthMe
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/auth/change-password:
    patch:
      tags:
        - auth
      summary: Change password of current user
      description: Change password of current user
      operationId: changePasswordMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  example: 123456789
                newPassword:
                  type: string
                  example: 123456789
              required:
                - newPassword
          application/xml:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                currentPassword:
                  type: string
              required:
                - newPassword
                - currentPassword
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                currentPassword:
                  type: string
              required:
                - newPassword
                - currentPassword
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/auth/forgot-password:
    post:
      tags:
        - auth
      summary: Forgot password of current user
      description: Forgot password of current user
      operationId: forgotPasswordMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@gmail.com
              required:
                - email
          application/xml:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
...
components:
  schemas:
...
    User:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          default: string
        phoneNumber:
          type: string
        address:
          type: string
        avatar:
          type: string
        city:
          type: string
        status:
          type: number
          default: 1
          enum: [0, 1]
        userType:
          type: number
          default: 3
          enum: [1, 2, 3]
        addresses:
          type: array
          items:
            type: object
            properties:
              firstName:
                type: string
              lastName:
                type: string
              middleName:
                type: string
              phoneNumber:
                type: string
              price:
                type: number
              discount:
                type: number
              product:
                type: string
                format: uuid
      required:
        - email
        - password
      xml:
        name: User
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT